Fix shadow for pre-21 devices and change colour when disabling app
