+++ app/build.gradle
@@ -33,9 +33,9 @@ tasks.withType(Test) {
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:23.3.0'
-    compile 'com.android.support:preference-v7:23.3.0'
-    compile 'com.android.support:preference-v14:23.3.0'
+    compile 'com.android.support:appcompat-v7:23.4.0'
+    compile 'com.android.support:preference-v7:23.4.0'
+    compile 'com.android.support:preference-v14:23.4.0'
     compile 'com.google.android.gms:play-services-analytics:8.4.0'
     testCompile 'junit:junit:4.12'
     testCompile 'org.mockito:mockito-all:1.10.19'
