+++ app/src/main/res/values/colors.xml
@@ -4,6 +4,7 @@
     <color name="primary_dark">#689f38</color>
     <color name="accent">#ffd740</color>
 
+    <color name="toolbar_app_disabled">#9e9e9e</color>
     <!--<color name="text_primary">#000000</color>-->
     <!--<color name="text_secondary">#8c000000</color>-->
     <!--<color name="text_primary_inverse">#ffffff</color>-->
